Various cleanups, etc
