+++ .github/workflows/unittest.yml
@@ -1,4 +1,4 @@
-name: Black and Test
+name: Tests
 
 on: [push, pull_request]
 
@@ -18,13 +18,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python setup.py install
-        pip install black
-        pip install pytest
-
-    - name: Black
-      run: black . --check
+        pip install .
 
     - name: Test with pytest
-      run: |
-        pytest
+      run:
+        python setup.py test
