+++ setup.cfg
@@ -1,0 +1,5 @@
+[aliases]
+test=pytest
+
+[tool:pytest]
+python_files = tests/*.py
