+++ setup.py
@@ -31,6 +31,8 @@ setup(
     name="twitter-cleanup",
     packages=find_packages(),
     py_modules=["twitter_cleanup"],
+    setup_requires=['pytest-runner'],
+    tests_require=['pytest'],
     url="https://github.com/cuducos/twitter-cleanup",
     version="0.0.6",
     zip_safe=False,
